# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Timetracker API Server"
        }
    ],
    "paths": {
...
        "/roles": {
            "get": {
                "tags": [
                    "Users roles"
                ],
                "summary": "Get list of users roles",
                "description": "Returns list of users roles",
                "operationId": "getUserRolesList",
                "parameters": [
                    {
                        "name": "Page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Per page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "role": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "role": "user",
                                                    "created_at": "2021-04-22T16:06:27.000000Z",
                                                    "updated_at": "2021-04-22T16:06:27.000000Z",
                                                    "deleted_at": null
                                                },
                                                {
                                                    "id": 2,
                                                    "role": "admin",
                                                    "created_at": "2021-04-22T16:06:27.000000Z",
                                                    "updated_at": "2021-04-22T16:06:27.000000Z",
                                                    "deleted_at": null
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users roles"
                ],
                "summary": "Store new user role",
                "description": "Returns user role data",
                "operationId": "storeUserRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "role": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 1,
                                        "role": "user",
                                        "created_at": "2021-04-22T16:06:27.000000Z",
                                        "updated_at": "2021-04-22T16:06:27.000000Z",
                                        "deleted_at": null
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Users roles"
                ],
                "summary": "Get user role information",
                "description": "Returns user role data",
                "operationId": "getUserRoleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UserRole id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "role": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 2,
                                        "role": "admin",
                                        "created_at": "2021-04-22T16:06:27.000000Z",
                                        "updated_at": "2021-04-22T16:06:27.000000Z",
                                        "deleted_at": null
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users roles"
                ],
                "summary": "Update existing user role",
                "description": "Returns updated user role data",
                "operationId": "updateUserRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UserRole id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "role": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 2,
                                        "role": "admin",
                                        "created_at": "2021-04-22T16:06:27.000000Z",
                                        "updated_at": "2021-04-22T16:06:27.000000Z",
                                        "deleted_at": null
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users roles"
                ],
                "summary": "Delete existing user role",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteUserRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UserRole id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}